Renamed async/sync for tasks/threads
