+++ concurrency/src/lib.rs
@@ -1,5 +1,5 @@
 //! λ-kit concurrency
 //! Some basic traits and structs to implement À-la-Erlang concurrent code.
 
-pub mod r#async;
-pub mod sync;
+pub mod tasks;
+pub mod threads;
