--- concurrency/src/sync/error.rs
@@ -1,11 +1,0 @@
-#[derive(Debug)]
-pub enum GenServerError {
-    CallbackError,
-    ServerError,
-}
-
-impl<T> From<spawned_rt::sync::mpsc::SendError<T>> for GenServerError {
-    fn from(_value: spawned_rt::sync::mpsc::SendError<T>) -> Self {
-        Self::ServerError
-    }
-}
