+++ concurrency/src/tasks/time.rs
@@ -1,8 +1,8 @@
 use std::time::Duration;
 
-use spawned_rt::sync::{self as rt, mpsc::Sender, JoinHandle};
+use spawned_rt::tasks::{self as rt, mpsc::Sender, JoinHandle};
 
-use super::gen_server::{GenServer, GenServerInMsg};
+use super::{GenServer, GenServerInMsg};
 
 // Sends a message after a given period to the specified GenServer. The task terminates
 // once the send has completed
@@ -14,8 +14,8 @@ pub fn send_after<T>(
 where
     T: GenServer + 'static,
 {
-    rt::spawn(move || {
-        rt::sleep(period);
+    rt::spawn(async move {
+        rt::sleep(period).await;
         let _ = tx.send(GenServerInMsg::Cast { message });
     })
 }
