+++ concurrency/src/threads/process.rs
@@ -1,7 +1,7 @@
 //! Process trait and struct to create a process abstraction similar to Erlang processes.
 //! See examples/ping_pong for a usage example.
 
-use spawned_rt::sync::{self as rt, mpsc, JoinHandle};
+use spawned_rt::threads::{self as rt, mpsc, JoinHandle};
 
 #[derive(Debug)]
 pub struct ProcessInfo<T> {
