+++ concurrency/src/threads/time.rs
@@ -1,8 +1,8 @@
 use std::time::Duration;
 
-use spawned_rt::r#async::{self as rt, mpsc::Sender, JoinHandle};
+use spawned_rt::threads::{self as rt, mpsc::Sender, JoinHandle};
 
-use super::{GenServer, GenServerInMsg};
+use super::gen_server::{GenServer, GenServerInMsg};
 
 // Sends a message after a given period to the specified GenServer. The task terminates
 // once the send has completed
@@ -14,8 +14,8 @@ pub fn send_after<T>(
 where
     T: GenServer + 'static,
 {
-    rt::spawn(async move {
-        rt::sleep(period).await;
+    rt::spawn(move || {
+        rt::sleep(period);
         let _ = tx.send(GenServerInMsg::Cast { message });
     })
 }
