+++ examples/bank/src/main.rs
@@ -26,8 +26,8 @@ use std::collections::HashMap;
 
 use messages::{BankError, BankOutMessage};
 use server::Bank;
-use spawned_concurrency::r#async::GenServer as _;
-use spawned_rt::r#async as rt;
+use spawned_concurrency::tasks::GenServer as _;
+use spawned_rt::tasks as rt;
 
 fn main() {
     rt::run(async {
