+++ examples/name_server_with_error/src/main.rs
@@ -18,8 +18,8 @@ use std::collections::HashMap;
 
 use messages::NameServerOutMessage;
 use server::NameServer;
-use spawned_concurrency::r#async::GenServer as _;
-use spawned_rt::r#async as rt;
+use spawned_concurrency::tasks::GenServer as _;
+use spawned_rt::tasks as rt;
 
 fn main() {
     rt::run(async {
