+++ examples/ping_pong_threads/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ping_pong_sync"
+name = "ping_pong_threads"
 version = "0.1.0"
 edition = "2021"
 
@@ -9,5 +9,5 @@ spawned-concurrency = { workspace = true }
 tracing = { workspace = true }
 
 [[bin]]
-name = "ping_pong_sync"
+name = "ping_pong_threads"
 path = "src/main.rs"
