+++ examples/updater_threads/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "updater_sync"
+name = "updater_threads"
 version = "0.1.0"
 edition = "2021"
 
@@ -11,5 +11,5 @@ reqwest = { version = "0.11", features = ["blocking"] }
 futures = "0.3.1"
 
 [[bin]]
-name = "updater_sync"
+name = "updater_threads"
 path = "src/main.rs"
