+++ examples/updater_threads/src/messages.rs
