+++ rt/src/tasks/mod.rs
@@ -11,10 +11,10 @@ mod tokio;
 
 use crate::tracing::init_tracing;
 
-pub use crate::r#async::tokio::mpsc;
-pub use crate::r#async::tokio::oneshot;
-pub use crate::r#async::tokio::sleep;
-pub use crate::r#async::tokio::{spawn, JoinHandle, Runtime};
+pub use crate::tasks::tokio::mpsc;
+pub use crate::tasks::tokio::oneshot;
+pub use crate::tasks::tokio::sleep;
+pub use crate::tasks::tokio::{spawn, JoinHandle, Runtime};
 use std::future::Future;
 
 pub fn run<F: Future>(future: F) -> F::Output {
